fix: Resolved db connection issues

+++ backend/src/main/java/com/swappable/backend/user/UserService.java
@@ -6,26 +6,21 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class UserService {
-
     private final UserRepository userRepository;
 
-    // Constructor links the repository so the service will use it
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
 
     @Transactional
-    public void updateUserProfile(Long userId, UserUpdateDto updateDto) {
-        // Finding the user or gives error if they don't exist
+    public void updateUserProfile(Integer userId, UserUpdateDto updateDto) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
-        // Updating the fields
         user.setBio(updateDto.bio());
         user.setLocation(updateDto.location());
         user.setDisplayName(updateDto.displayName());
 
-        // Saving changes to go back the database
         userRepository.save(user);
     }
 }
